direct-tree-eval: refactor: decouple ErgoTreeEvaluator
