+++ sigmastate/src/test/scala/sigmastate/helpers/SigmaPPrint.scala
@@ -17,8 +17,7 @@ import sigmastate.lang.SigmaTyper
 import sigmastate.lang.Terms.MethodCall
 import sigmastate.serialization.GroupElementSerializer
 import sigmastate.utxo.SelectField
-import sigmastate.interpreter.ErgoTreeEvaluator.{CompanionDesc, MethodDesc}
-import sigmastate.interpreter.{ErgoTreeEvaluator, FixedCostItem}
+import sigmastate.interpreter.{CompanionDesc, ErgoTreeEvaluator, FixedCostItem, MethodDesc}
 import special.collection.Coll
 import special.sigma.GroupElement
